✨(frontend) update course form
Transform the course form by separating it into several tabs
