+++ src/frontend/admin/src/components/templates/courses/form/translations.ts
@@ -1,6 +1,12 @@
 import { defineMessages } from "react-intl";
 
 export const courseFormMessages = defineMessages({
+  generalSectionInfo: {
+    id: "components.templates.course.form.translations.generalSectionInfo",
+    defaultMessage:
+      "In this section, you can configure your course by filling in all the required fields.",
+    description: "Text for the general section info",
+  },
   generalSubtitle: {
     id: "components.templates.course.form.translations.generalSubtitle",
     defaultMessage: "Main information",
@@ -36,9 +42,15 @@ export const courseFormMessages = defineMessages({
     defaultMessage: "Cover",
     description: "Label for the cover field",
   },
+  targetCourseRunAlertInfo: {
+    id: "components.templates.course.form.translations.targetCourseRunAlertInfo",
+    defaultMessage:
+      "In this section, you can view, add, modify or delete the course runs linked to this course",
+    description: "Title for the course run section info",
+  },
   targetCourseRunTitle: {
     id: "components.templates.course.form.translations.targetCourseRunTitle",
-    defaultMessage: "Course runs of this course",
+    defaultMessage: "Relation to course runs",
     description: "Title for the course run section inside the course form",
   },
   addCourseRunModalTitle: {
@@ -61,9 +73,41 @@ export const courseFormMessages = defineMessages({
     defaultMessage: "Add relation",
     description: "Label for the product relation subheader form",
   },
+  membersSectionInfo: {
+    id: "components.templates.course.form.translations.membersSectionInfo",
+    defaultMessage:
+      "In this section you can manage access to this course. You can also edit roles easily. Please note, you must have at least one owner for a course.",
+    description: "Text for the member info section",
+  },
   membersSectionTitle: {
     id: "components.templates.course.form.translations.membersSectionTitle",
     defaultMessage: "Course members",
     description: "Title for the member section",
+  },
+  generalTabTitle: {
+    id: "components.templates.course.form.translations.generalTabTitle",
+    defaultMessage: "General",
+    description: "Title for the general tab",
+  },
+  membersTabTitle: {
+    id: "components.templates.course.form.translations.membersTabTitle",
+    defaultMessage: "Members",
+    description: "Title for the members tab",
+  },
+  courseRunsTabTitle: {
+    id: "components.templates.course.form.translations.courseRunsTabTitle",
+    defaultMessage: "Course runs",
+    description: "Title for the course runs tab",
+  },
+  productsTabInfo: {
+    id: "components.templates.course.form.translations.productsTabTitle",
+    description: "Help text to explain what this section is about",
+    defaultMessage:
+      "In this section, you can choose the products that will be displayed on this course.",
+  },
+  productsTabTitle: {
+    id: "components.templates.course.form.translations.productsTabTitle",
+    defaultMessage: "Products",
+    description: "Title for the products tab",
   },
 });
